requirements.txt
imapclient
feedgen
Flask
requests
python-dotenv
APScheduler